requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
llama-index
llama-index-vector-stores-pinecone
llama-index-embeddings-openai
llama-index-llms-openai
pinecone
openai
pypdf2
python-docx
pandas
python-slugify
sentence-transformers
python-multipart
requests

